remove legacy models file
