+++ backend/api/health.py
@@ -1,0 +1,17 @@
+from fastapi import APIRouter
+
+from ..game import DICTIONARY
+
+router = APIRouter()
+
+
+@router.get("/health")
+def health() -> dict[str, str]:
+    """Health check endpoint."""
+    return {"status": "ok"}
+
+
+@router.get("/validate")
+def validate(word: str) -> dict[str, bool]:
+    """Validate a word against the ODS8 dictionary."""
+    return {"valid": word.upper() in DICTIONARY}
