+++ backend/models/__init__.py
@@ -1,0 +1,14 @@
+from .base import Base, TimestampMixin
+from .user import User, OAuthAccount
+from .game import Game, GamePlayer, PlacedTile, Word
+
+__all__ = [
+    "Base",
+    "TimestampMixin",
+    "User",
+    "OAuthAccount",
+    "Game",
+    "GamePlayer",
+    "PlacedTile",
+    "Word",
+]
