+++ backend/tests/bot/test_bot_turn.py
@@ -9,7 +9,7 @@ sys.path.append(str(pathlib.Path(__file__).resolve().parents[2]))
 os.environ["DATABASE_URL"] = "sqlite:///./test.db"
 
 from backend.database import Base, SessionLocal, engine  # type: ignore
-from backend.main import (
+from backend.api.games import (
     CreateGameRequest,
     JoinGameRequest,
     MoveRequest,
@@ -19,7 +19,7 @@ from backend.main import (
     start_game,
     get_game_state,
     _maybe_play_bot,
-)  # type: ignore
+)
 from backend import models  # type: ignore
 
 Base.metadata.drop_all(bind=engine)
@@ -39,19 +39,22 @@ def _setup_bot_game():
     with SessionLocal() as db:
         start_data = start_game(game_id, seed=0, db=db)
     rack1 = next(p["rack"] for p in start_data["players"] if p["player_id"] == p1)
-    return game_id, p1, p2, rack1
+    rack_bot = next(p["rack"] for p in start_data["players"] if p["player_id"] == p2)
+    return game_id, p1, p2, rack1, rack_bot
 
 
 def test_bot_move_triggered_after_player_turn():
-    game_id, p1, _bot_id, rack1 = _setup_bot_game()
+    game_id, p1, _bot_id, rack1, rack_bot = _setup_bot_game()
     placements = [
         {"row": 7, "col": 7, "letter": rack1[1], "blank": False},
         {"row": 7, "col": 8, "letter": rack1[3], "blank": False},
         {"row": 7, "col": 9, "letter": rack1[2], "blank": False},
     ]
+    bot_letter = rack_bot[0].upper()
     with patch(
-        "backend.main.game_module.bot_turn", return_value=([(7, 10, "A", False)], 1)
-    ), patch("backend.main.place_tiles", return_value=1):
+        "backend.api.games.game_module.bot_turn",
+        return_value=([(7, 10, bot_letter, False)], 1),
+    ), patch("backend.api.games.place_tiles", return_value=1):
         with SessionLocal() as db:
             res = play_move(game_id, MoveRequest(player_id=p1, placements=placements), db=db)
     assert "bot_move" in res
@@ -62,14 +65,14 @@ def test_bot_move_triggered_after_player_turn():
 
 
 def test_invalid_bot_move_keeps_turn():
-    game_id, p1, bot_id, rack1 = _setup_bot_game()
+    game_id, p1, bot_id, rack1, _rack_bot = _setup_bot_game()
     placements = [
         {"row": 7, "col": 7, "letter": rack1[1], "blank": False},
         {"row": 7, "col": 8, "letter": rack1[3], "blank": False},
         {"row": 7, "col": 9, "letter": rack1[2], "blank": False},
     ]
     with patch(
-        "backend.main.game_module.bot_turn", return_value=([(0, 0, "A", False)], 1)
+        "backend.api.games.game_module.bot_turn", return_value=([(0, 0, "A", False)], 1)
     ):
         with SessionLocal() as db:
             res = play_move(game_id, MoveRequest(player_id=p1, placements=placements), db=db)
@@ -78,19 +81,20 @@ def test_invalid_bot_move_keeps_turn():
 
 
 def test_state_fetch_triggers_pending_bot_move():
-    game_id, p1, bot_id, rack1 = _setup_bot_game()
+    game_id, p1, bot_id, rack1, rack_bot = _setup_bot_game()
     placements = [
         {"row": 7, "col": 7, "letter": rack1[1], "blank": False},
         {"row": 7, "col": 8, "letter": rack1[3], "blank": False},
         {"row": 7, "col": 9, "letter": rack1[2], "blank": False},
     ]
-    with patch("backend.main.game_module.bot_turn", return_value=None):
+    with patch("backend.api.games.game_module.bot_turn", return_value=None):
         with SessionLocal() as db:
             res = play_move(game_id, MoveRequest(player_id=p1, placements=placements), db=db)
     assert res["next_player_id"] == bot_id
     with patch(
-        "backend.main.game_module.bot_turn", return_value=([(7, 10, "A", False)], 1)
-    ), patch("backend.main.place_tiles", return_value=1):
+        "backend.api.games.game_module.bot_turn",
+        return_value=([(7, 10, rack_bot[0].upper(), False)], 1),
+    ), patch("backend.api.games.place_tiles", return_value=1):
         with SessionLocal() as db:
             state = get_game_state(game_id, player_id=p1, db=db)
     assert state.next_player_id == p1
@@ -103,7 +107,7 @@ def test_bot_reloads_board_before_playing():
     This simulates a scenario where the in-memory board is cleared (e.g. after a
     server restart) before the bot attempts its move.
     """
-    game_id, p1, bot_id, rack1 = _setup_bot_game()
+    game_id, p1, bot_id, rack1, rack_bot = _setup_bot_game()
     placements = [
         {"row": 7, "col": 7, "letter": rack1[1], "blank": False},
         {"row": 7, "col": 8, "letter": rack1[3], "blank": False},
@@ -114,7 +118,7 @@ def test_bot_reloads_board_before_playing():
         players = db.query(models.GamePlayer).filter_by(game_id=game_id).all()
         return players, None, 0
 
-    with patch("backend.main._maybe_play_bot", side_effect=no_bot):
+    with patch("backend.api.games._maybe_play_bot", side_effect=no_bot):
         with SessionLocal() as db:
             play_move(game_id, MoveRequest(player_id=p1, placements=placements), db=db)
 
@@ -126,13 +130,13 @@ def test_bot_reloads_board_before_playing():
     def fake_bot_turn(rack):
         # Board should be reloaded with the player's move before bot_turn is called
         assert game_module.board[7][7] is not None
-        return ([(7, 10, "A", False)], 1)
+        return ([(7, 10, rack_bot[0].upper(), False)], 1)
 
-    with patch("backend.main.game_module.bot_turn", side_effect=fake_bot_turn), patch(
-        "backend.main.place_tiles", return_value=1
+    with patch("backend.api.games.game_module.bot_turn", side_effect=fake_bot_turn), patch(
+        "backend.api.games.place_tiles", return_value=1
     ):
         with SessionLocal() as db:
             game = db.get(models.Game, game_id)
             _players, bot_move, _score = _maybe_play_bot(game_id, game, db)
 
-    assert bot_move == [(7, 10, "A", False)]
+    assert bot_move == [(7, 10, rack_bot[0].upper(), False)]
