+++ backend/tests/bot/test_bot_validation.py
