--- backend/tests/test_user_connexion.py
@@ -1,4 +1,0 @@
-def create_user()->None:
-
-def connexion_user()->None:
-
